requirements.txt
# Core cryptographic libraries
PyCryptodome>=3.18.0
cryptography>=41.0.0

# GUI framework
# tkinter is usually included with Python

# Mathematical operations
sympy>=1.12

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.0.0
black>=23.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0